Add full prepare download and comment individual download/delete video

+++ util/download.py
@@ -3,11 +3,10 @@ import os
 
 def download_video(url, file_path, file_name, file_extension):
     try:
-        try:
-            urllib.request.urlretrieve(url, '%s/%s.%s'%(file_path, file_name, file_extension))
-        except:
+        if not os.path.exists(file_path):
             os.mkdir(file_path)
+        if not os.path.exists('%s/%s.%s'%(file_path, file_name, file_extension)):
             urllib.request.urlretrieve(url, '%s/%s.%s'%(file_path, file_name, file_extension))
     except:
-        print('Cannot write because file "%s/%s.%s" not found'%(file_path, file_name, file_extension))
+        print('Cannot download file "%s/%s.%s".'%(file_path, file_name, file_extension))
         raise Exception
